# spec/0.1/facilitator-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: v402 Facilitator API
  version: "0.1"
  description: >-
    Payment-infrastructure service between customer and API operator:
    signature verification, balance state, deposit watching, blocklist
    enforcement. Contract for any implementation (middleware calling a
    facilitator, or a facilitator serving middlewares). See
    facilitator-api.md for normative prose and the error catalog.
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/
security: []
...
paths:
...
  /v1/balance:
    get:
      tags: [public]
      operationId: balance
      summary: Signature-authenticated balance query
      description: >-
        Authenticates via a VerusID signature over the domain-separated
        `v402-balance-query/0.1` canonical payload (see canonical-payload.md).
        Ownership is evaluated against the LATEST identity state
        (checklatest) — rotated/revoked keys are rejected. Replay-protected
        like payments (repeat → 409 with `previousStatus`).
      parameters:
        - name: X-V402-Payer
          in: header
          required: true
          schema: { type: string }
        - name: X-V402-Request-Id
          in: header
          required: true
          schema: { type: string }
          description: Fresh ULID per query
        - name: X-V402-Issued-At
          in: header
          required: true
          schema: { type: string }
          description: Unix seconds, ±300 s acceptance window
        - name: X-V402-Signature
          in: header
          required: true
          schema: { type: string }
          description: Base64 signature over the balance-query canonical payload
      responses:
        "200":
          description: Balance state (zeros for identities without an account)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceResponse" }
        default: { $ref: "#/components/responses/Error" }
...
  /admin/simulate-deposit:
    post:
      tags: [admin]
      operationId: simulateDeposit
      summary: Credit a simulated deposit (simulated watcher mode only)
      description: >-
        409 `not-simulated` when the watcher runs in real mode. Credits
        instantly, records `origin: "simulated"`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SimulateDepositRequest" }
      responses:
        "201":
          description: Deposit credited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCreditResponse" }
        "401": { $ref: "#/components/responses/AuthError" }
        default: { $ref: "#/components/responses/Error" }
  /admin/credit:
    post:
      tags: [admin]
      operationId: adminCredit
      summary: Manual support credit (any mode)
      description: >-
        Recorded as simulated origin — excluded from on-chain crosschecks;
        document the support process before crediting REAL missed deposits
        this way.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCreditRequest" }
      responses:
        "201":
          description: Credit booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCreditResponse" }
        "401": { $ref: "#/components/responses/AuthError" }
        default: { $ref: "#/components/responses/Error" }
  /admin/reconcile:
    post:
      tags: [admin]
      operationId: reconcile
      summary: On-demand ledger-invariant check
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Reconciliation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReconcileResponse" }
        "401": { $ref: "#/components/responses/AuthError" }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        username = middleware identifier (free-form, logged), password = the
        operator-provisioned middleware token. Fail closed: a facilitator
        with no token configured rejects all payment calls.
    bearerAuth:
      type: http
      scheme: bearer
      description: Operator admin token; fail closed when unconfigured.
  responses:
    Error:
      description: >-
        Error envelope. `error.code` values are stable identifiers — see
        facilitator-api.md § Error catalog for the full list and the HTTP
        status per class (400 request errors, 402 payment, 403 blocked,
        404 unknown, 409 replay/state, 503 verify-unavailable). Clients MUST
        tolerate unknown codes.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    AuthError:
      description: >-
        Authentication failure (missing/invalid Basic or Bearer credentials).
        NOTE — reference-implementation reality: auth failures use the
        framework's default error shape, NOT the v402 error envelope. Clients
        should treat any 401 as "fix credentials" without parsing the body.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FrameworkError" }
  schemas:
    SatsString:
      type: string
      pattern: "^-?[0-9]+$"
      description: >-
        Exact satoshi amount as a decimal string (never a JSON number).
        Negative only where explicitly allowed (late-commit balanceAfterSats).
    HumanAmount:
      type: string
      pattern: "^[0-9]+(\\.[0-9]+)?$"
      description: Human decimal amount in minimal form ("0.001", never padded)
...
    UnixSeconds:
      type: integer
      minimum: 0
    FrameworkError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message: { type: string }
        error: { type: string }
      additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable identifier from the error catalog (open set)
            message: { type: string }
            details:
              type: object
              description: >-
                Code-specific extras, e.g. `previousStatus` (replay),
                `balanceSats`/`requiredSats`/`depositAddress`
                (insufficient-balance), `retryAfterSec` (verify-unavailable),
                current `accepts` (price-mismatch).
      additionalProperties: true
...
    BalanceResponse:
      type: object
      required:
        [identity, balance, reserved, available, balanceSats, reservedSats, availableSats]
      properties:
        identity:
          type: string
          description: Normalized identity key
        balance: { $ref: "#/components/schemas/HumanAmount" }
        reserved: { $ref: "#/components/schemas/HumanAmount" }
        available: { $ref: "#/components/schemas/HumanAmount" }
        pending:
          allOf: [{ $ref: "#/components/schemas/HumanAmount" }]
          description: >-
            Detected but not yet credited deposits (confirmations
            accumulating). Additive 2026-07-14; absent on older facilitators.
        balanceSats: { $ref: "#/components/schemas/SatsString" }
        reservedSats: { $ref: "#/components/schemas/SatsString" }
        availableSats: { $ref: "#/components/schemas/SatsString" }
        pendingSats: { $ref: "#/components/schemas/SatsString" }
        firstDepositAt: { $ref: "#/components/schemas/UnixSeconds" }
        lastRequestAt: { $ref: "#/components/schemas/UnixSeconds" }
      additionalProperties: true
...
    SimulateDepositRequest:
      type: object
      required: [identity, amount, operator]
      properties:
        identity: { type: string }
        amount: { $ref: "#/components/schemas/HumanAmount" }
        operator:
          type: string
          minLength: 1
          maxLength: 100
          description: Operator attribution — who injected this deposit (audited).
        txid: { type: string }
        note: { type: string, maxLength: 500 }
    AdminCreditRequest:
      type: object
      required: [identity, amount, operator]
      properties:
        identity: { type: string }
        amount: { $ref: "#/components/schemas/HumanAmount" }
        operator:
          type: string
          minLength: 1
          maxLength: 100
          description: Operator attribution — who booked this credit (audited).
        txid: { type: string }
        note: { type: string, maxLength: 500 }
    AdminCreditResponse:
      type: object
      required: [ok, deposit, identity, balanceAfterSats]
      properties:
        ok: { const: true }
        deposit:
          type: object
          required: [id, txid, origin]
          properties:
            id: { type: integer }
            txid: { type: string }
            origin: { type: string, enum: [real, simulated] }
          additionalProperties: true
        identity:
          type: string
          description: Normalized identity key the credit was booked to
        balanceAfterSats: { $ref: "#/components/schemas/SatsString" }
      additionalProperties: true
    ReconcileResponse:
      type: object
      required: [ok, runAt, identitiesChecked, mismatches, detail, onChain, durationMs]
      properties:
        ok: { const: true }
        runAt: { $ref: "#/components/schemas/UnixSeconds" }
        identitiesChecked: { type: integer }
        mismatches: { type: integer }
        detail:
          type: array
          items: { type: object, additionalProperties: true }
        onChain:
          oneOf:
            - type: object
              required: [available]
              properties:
                available: { const: false }
              additionalProperties: false
            - type: object
              required: [available, creditedDepositSats, chainBalanceSats]
              properties:
                available: { const: true }
                creditedDepositSats: { $ref: "#/components/schemas/SatsString" }
                chainBalanceSats: { $ref: "#/components/schemas/SatsString" }
              additionalProperties: false
          description: Advisory on-chain crosscheck — NOT counted as a mismatch
        durationMs: { type: integer }
      additionalProperties: true